+++ src/main/java/seedu/taskmaster/model/Model.java
@@ -151,14 +151,14 @@ public interface Model {
      * Marks the attendance of the given student {@code target} with the given {@code attendanceType}.
      * The student must exist in the student list.
      */
-    void scoreStudent(StudentRecord target, int score);
+    void scoreStudent(StudentRecord target, double score);
 
-    void scoreStudentWithNusnetId(NusnetId nusnetId, int score);
+    void scoreStudentWithNusnetId(NusnetId nusnetId, double score);
 
     /**
      * Marks the attendances of all {@code students} with the given {@code attendanceType}
      */
-    void scoreAllStudents(List<StudentRecord> students, int score);
+    void scoreAllStudents(List<StudentRecord> students, double score);
 
     /**
      * Updates the corresponding attendance statuses with the Attendances in the given list.
