+++ src/main/java/seedu/taskmaster/model/record/StudentRecordList.java
@@ -21,12 +21,12 @@ public interface StudentRecordList extends Iterable<StudentRecord> {
     /**
      * Updates participation score of a student represented by their {@code nusnetId} to {@code score}.
      */
-    void scoreStudentParticipation(NusnetId nusnetId, int score);
+    void scoreStudentParticipation(NusnetId nusnetId, double score);
 
     /**
      * Updates participation score of all students in the list of {@code nusnetIds} with {@code attendanceType}.
      */
-    void scoreAllParticipation(List<NusnetId> nusnetIds, int score);
+    void scoreAllParticipation(List<NusnetId> nusnetIds, double score);
 
 
     void setStudentRecords(StudentRecordListManager replacement);
